requirements.txt
gradio
scikit-learn
pandas
numpy
matplotlib
joblib
pyyaml
optuna
catboost
xgboost
lightgbm
tensorflow  
pyarrow
fastparquet 
ipykernel
seaborn
tqdm
statsmodels
shap
nbformat
ipywidgets
gradio
tabulate